Done with win version after stuff disappeared

+++ pydl-win/pydl/libraries/download_convert.py
@@ -1,7 +1,7 @@
 # type: ignore
 from ffmpeg import FFmpeg
 import os
-import sys
+from pathlib import Path
 from time import sleep
 
 # rich
@@ -21,6 +21,8 @@ from libraries import get_metadata
 from libraries import notifier
 from libraries import find_stream
 from libraries import queue_management
+from libraries import read_conf
+from libraries import upload as uploader
 
 from pydl import console
 
@@ -29,13 +31,19 @@ tn_as_cover = None
 # 3 options for tagging: normal, experimental, off
 tagging = "normal"
 
+# complex file structure is if we place the song in the artist/album/ directory
+complex_file_structure = True
+
 # the path where the songs are downloaded to
 dl_path = os.getcwd()
 
-# album name is needed by manage playlist for notifying the user correctly
+output_dir = None
+
+# album name is needed by manfage playlist for notifying the user correctly
 album_name = None
 
 status = True
+upload = None
 
 
 def download(url):
@@ -46,15 +54,20 @@ def download(url):
     return song
 
 
-def convert(song, title):
+def convert(song, output_dir, title):
     # set the new filename with the mp3 extension
-    filename = song.split(".")[0] + ".mp3"
+    filename = Path(song).stem + ".mp3"
+    filename = os.path.join(output_dir, filename)
 
     # if we would not check this ffmpeg would hast ot itself. That would mess up the status
     if os.path.isfile(filename):
-        # send a notification to the user that input is required
-        notifier.error(title)
+        if not read_conf.main()["Flags"]["skip-existing"]:
+            # send a notification to the user that input is required
+            notifier.error(title)
         if not status:
+            if read_conf.main()["Flags"]["skip-existing"]:
+                cleanup(song)
+                return filename
             queue_management.error()
             console.print(
                 f"[bold red]File {filename} already exists, remove or rename the file!"
@@ -70,21 +83,15 @@ def convert(song, title):
                     f"[bold green]Converting {title}...[/bold green]", spinner="dots"
                 ):
                     # add the parameters to the ffmpeg command
-                    ffmpeg = (FFmpeg()
-                          .option("y")
-                          .input(song)
-                          .output(filename))
-               
+                    ffmpeg = FFmpeg().option("y").input(song).output(filename)
+
                     # perform the conversion
                     ffmpeg.execute()
             else:
                 queue_management.set_status_text(f"Converting {title}...")
                 # add the parameters to the ffmpeg command
-                ffmpeg = (FFmpeg()
-                          .option("y")
-                          .input(song)
-                          .output(filename))
-               
+                ffmpeg = FFmpeg().option("y").input(song).output(filename)
+
                 # perform the conversion
                 ffmpeg.execute()
                 queue_management.set_status_text("")
@@ -96,36 +103,35 @@ def convert(song, title):
                 f"[bold green]Converting {title}...[/bold green]", spinner="dots"
             ):
                 # add the parameters to the ffmpeg command
-                ffmpeg = (FFmpeg()
-                          .option("y")
-                          .input(song)
-                          .output(filename))
-               
+                ffmpeg = FFmpeg().option("y").input(song).output(filename)
+
                 # perform the conversion
                 ffmpeg.execute()
         else:
             queue_management.set_status_text(f"Converting {title}...")
-            ffmpeg = (FFmpeg()
-                          .option("y")
-                          .input(song)
-                          .output(filename))
-               
+            ffmpeg = FFmpeg().option("y").input(song).output(filename)
+
             # perform the conversion
             ffmpeg.execute()
             queue_management.set_status_text("")
-
     # return the new filename so we can work with it later
     return filename
 
 
 def cleanup(song):
-    os.remove(song)
+    if os.path.isfile(song):
+        os.remove(song)
 
 
 # these functions get called by the main script. only needed to make the code easier to understand
 def set_tagging(tagging_mode):
     global tagging
     tagging = tagging_mode
+
+
+def set_complex_structure(complex_structure):
+    global complex_file_structure
+    complex_file_structure = complex_structure
 
 
 def set_path(path):
@@ -135,7 +141,6 @@ def set_path(path):
 
 def main(url, tags_in, song_nr, playlist_title):
     global album_name
-
     tags_out = {
         "artist": None,
         "album": None,
@@ -150,7 +155,7 @@ def main(url, tags_in, song_nr, playlist_title):
 
     # get the title of the song, this is done in a seperate function beacuse pytube throws random errors sometimes.
     # to avoid this we use a try except block and a loop. so we just try again if the fetch fails
-    title = get_title(url)
+    title = get_metadata.get_title(url)
 
     # download the song in the highest quality available
     if status:
@@ -163,8 +168,9 @@ def main(url, tags_in, song_nr, playlist_title):
         downloaded_song = download(url)
         queue_management.set_status_text("")
 
+    output_dir = create_folder_structure(downloaded_song, playlist_title)
     # convert the song to mp3
-    converted_song = convert(downloaded_song, title)
+    converted_song = convert(downloaded_song, output_dir, title)
 
     # remove the webm file
     cleanup(downloaded_song)
@@ -182,7 +188,7 @@ def main(url, tags_in, song_nr, playlist_title):
             if tagging == "normal":
                 # set the different tags if they are not overwritten by the user
                 if tags_out["artist"] == None:
-                    tags_out["artist"] = get_artist(url)
+                    tags_out["artist"] = get_metadata.get_artist(url)
                 if tags_out["title"] == None:
                     tags_out["title"] = title
 
@@ -292,32 +298,33 @@ def main(url, tags_in, song_nr, playlist_title):
             cleanup(img)
         queue_management.set_status_text("")
 
-
-# get the uploader of the video and set as artist additionally remove any unnecessary text
-def get_artist(url):
-    if "Topic" in str(CH(YT(url).channel_url).channel_name):
-        _artist = str(CH(YT(url).channel_url).channel_name).removesuffix(" - Topic")
-    elif "VEVO" in str(CH(YT(url).channel_url).channel_name):
-        _artist = str(CH(YT(url).channel_url).channel_name).removesuffix("VEVO")
-    else:
-        _artist = str(CH(YT(url).channel_url).channel_name)
-
-    return _artist
-
-
-def get_title(url):
-    error = True
-    while error == True:
-        # get the title of the song
-        try:
-            title = pyt.YouTube(url).title
-            error = False
-        except PytubeError:
-            pass
-
-    return title
+    # upload the song if the user wants to
+    if upload:
+        uploader.upload_song(converted_song)
 
 
 def set_status(_status):
     global status
     status = _status
+
+
+def set_upload(_upload):
+    global upload
+    upload = _upload
+
+
+def create_folder_structure(song, album_name):
+    global output_dir
+    parent_dir = os.path.abspath(os.path.join(song, os.pardir))
+    if complex_file_structure:
+        artist = get_metadata.main(song)["artist"]
+        if not os.path.exists(os.path.join(parent_dir, artist)):
+            os.mkdir(os.path.join(parent_dir, artist))
+        if not os.path.exists(os.path.join(parent_dir, artist, album_name)):
+            os.mkdir(os.path.join(parent_dir, artist, album_name))
+        out_directory = os.path.join(parent_dir, artist, album_name)
+        output_dir = out_directory
+        return out_directory
+    else:
+        output_dir = parent_dir
+        return parent_dir
